Use `zcash_primitives` and `zcash_address` versions corrsponding to `zcashd 6.2.0`.
This crate is intended exclusively for parsing and migration of `zcashd`
`wallet.dat` data into the `zewif` format. In order to ensure maximal
compatibility with `zcashd` deserialization, we use the same dependency
versions as the Rust code in the `zcashd` crate. This will become
particularly important when it comes to parsing the Orchard note
commitment tree data.