Use `zcash_primitives` to perform parsing for protocol types.
This replaces locally-defined parsing logic with `zcashd`'s parsing
code. By using the same versions of `zcash_primitives` and `bridgetree`
as `zcashd` uses, we can ensure compatibility and gain access to the
protocol data that is needed for type-safe conversion into the `zewif`
formats.

This also moves the `parser` infrastructure here from the `zewif` crate.
The encoding handled by these parsers is `zcashd`-specific, and should
not be used by the wider ecosystem.