Fix incorrect parsing of `CWalletTx::mapSaplingNoteData`
In `zcashd`, `mapSaplingNoteData` is always parsed if the transaction
version supports sapling; the implementation in `wallet_tx` parsing
incorrectly parsed it instead as a value wrapped in `std::optional`.
This causes an off-by-one error in the remainder of `CWalletTx` parsing,
but that error only appears sometimes because the `0x00` and `0x01`
discriminants for optional values are also valid vector lengths, so
transactions that did not include a Sapling component would not have a
parse error here.